feat: drop _OUT suffix everywhere
For now, the only files that maintain the _OUT suffix are the "R_OUT_<section>_R=<value>" files generated by the R-factor calculation. However, the code is reworked such that we can decide to drop the '_OUT' part with minimal modifications.

Tests:
- backwards compatibility tests for cleanup.prerun_clean (was already implemented) and bookkeeper
- Refactored sections/test_cleanup.py into a package with submodules for all the (currently tested) public functions.
