+++ src/viperleed/calc/sections/initialization.py
@@ -131,10 +131,11 @@ def initialization(sl, rp, subdomain=False):
     tmpslab = copy.deepcopy(sl)
     tmpslab.sort_original()
     try:
-        poscar.write(tmpslab, filename='POSCAR_OUT', comments='all')
+        poscar.write(tmpslab, filename='POSCAR', comments='all')
     except Exception:
         logger.error("Exception occurred while writing new POSCAR")
         raise
+    rp.files_to_out.add('POSCAR')
     # generate POSCAR_oricell
     tmpslab.revert_unit_cell()
     try:
